+++ MSEstudiantes/MSEstudiantes/src/main/java/GestionEstudiantes/MSEstudiantes/service/EstudianteService.java
@@ -1,6 +1,7 @@
 package GestionEstudiantes.MSEstudiantes.service;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -8,21 +9,21 @@ import org.springframework.stereotype.Service;
 import GestionEstudiantes.MSEstudiantes.model.EstudianteModel;
 import GestionEstudiantes.MSEstudiantes.repository.IEstudianteRepository;
 
- @Service
- public class EstudianteService implements IEstudianteService {
-
+@Service
+public class EstudianteService implements IEstudianteService {
 
     @Autowired
-    IEstudianteRepository repository;
+    private IEstudianteRepository repository;
 
     @Override
     public List<EstudianteModel> findAll() {
-        return (List<EstudianteModel>)repository.findAll();
+        return repository.findAll();
     }
 
     @Override
-    public EstudianteModel findById(Integer id) {
-        return (EstudianteModel)repository.findById(id).get();
+    public EstudianteModel findById(Long id) {
+        Optional<EstudianteModel> estudianteOpt = repository.findById(id);
+        return estudianteOpt.orElse(null);
     }
 
     @Override
@@ -31,17 +32,32 @@ import GestionEstudiantes.MSEstudiantes.repository.IEstudianteRepository;
     }
 
     @Override
-    public EstudianteModel update(EstudianteModel model) {
-        return repository.save(model);
+    public EstudianteModel update(Long id, EstudianteModel model) {
+        Optional<EstudianteModel> optionalEstudiante = repository.findById(id);
+        if (optionalEstudiante.isPresent()) {
+            EstudianteModel estudiante = optionalEstudiante.get();
+            // Actualiza los campos necesarios
+            estudiante.setDni(model.getDni());
+            estudiante.setNombre(model.getNombre());
+            estudiante.setApe_paterno(model.getApe_paterno());
+            estudiante.setApe_materno(model.getApe_materno());
+            estudiante.setEdad(model.getEdad());
+            estudiante.setSexo(model.getSexo());
+            // Si tienes otros campos, actualízalos aquí
+
+            return repository.save(estudiante);
+        } else {
+            return null; // O maneja el caso de no encontrado según tu lógica
+        }
     }
 
     @Override
-    public Boolean delete(Integer id) {
-        repository.deleteById(id);
-        return true;
+    public Boolean delete(Long id) {
+        if (repository.existsById(id)) {
+            repository.deleteById(id);
+            return true;
+        } else {
+            return false;
+        }
     }
-    
-}   
-
-    
-
+}
